web/requirements.txt
requests
imap-tools
# pycrypto
pycryptodome==3.20.0
pybit==5.6.2
websocket-client
django==4.2.11
celery==5.3.6
django-celery-results==2.5.1
beautifulsoup4==4.12.3
gunicorn
psycopg2-binary
Redis==5.0.3
django-cors-headers
Pillow
# new
djangorestframework
drf-yasg
django-cors-headers